+++ blog/comments/models.py
@@ -5,7 +5,8 @@ class Comment(models.Model):
     text = models.TextField('Текст')
     article_id = models.IntegerField('Идентификатор статьи')
     parent_id = models.IntegerField('Идентификатор родительского комментария')
-    date = models.IntegerField('Дата публикации')
+    date = models.DecimalField(
+        'Дата публикации', max_digits=15, decimal_places=0)
 
     class Meta:
         verbose_name = 'Комментарий'
